fix paakage name
