+++ src/main/java/new_ladder/view/LadderFormView.java
@@ -1,4 +1,4 @@
-package Ladder.view;
+package new_ladder.view;
 
 import java.util.Scanner;
