--- src/test/java/Ladder/LadderGameTest.java
@@ -1,12 +1,0 @@
-package Ladder;
-
-import org.junit.Test;
-
-public class LadderGameTest {
-
-	@Test
-	public void test() {
-		
-	}
-
-}
